% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comprehend_operations.R
\name{comprehend_describe_document_classification_job}
\alias{comprehend_describe_document_classification_job}
\title{Gets the properties associated with a document classification job}
\usage{
comprehend_describe_document_classification_job(JobId)
}
\arguments{
\item{JobId}{[required] The identifier that Amazon Comprehend generated for the job. The
operation returns this identifier in its response.}
}
\description{
Gets the properties associated with a document classification job. Use
this operation to get the status of a classification job.
}
\section{Request syntax}{
\preformatted{comprehend$describe_document_classification_job(
  JobId = "string"
)
}
}

\keyword{internal}
