% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lexmodelbuildingservice_operations.R
\name{lexmodelbuildingservice_get_builtin_intent}
\alias{lexmodelbuildingservice_get_builtin_intent}
\title{Returns information about a built-in intent}
\usage{
lexmodelbuildingservice_get_builtin_intent(signature)
}
\arguments{
\item{signature}{[required] The unique identifier for a built-in intent. To find the signature for
an intent, see \href{https://developer.amazon.com/public/solutions/alexa/alexa-skills-kit/docs/built-in-intent-ref/standard-intents}{Standard Built-in Intents}
in the \emph{Alexa Skills Kit}.}
}
\description{
Returns information about a built-in intent.
}
\details{
This operation requires permission for the \code{lex:GetBuiltinIntent}
action.
}
\section{Request syntax}{
\preformatted{lexmodelbuildingservice$get_builtin_intent(
  signature = "string"
)
}
}

\keyword{internal}
