% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lexmodelbuildingservice_operations.R
\name{lexmodelbuildingservice_get_import}
\alias{lexmodelbuildingservice_get_import}
\title{Gets information about an import job started with the StartImport
operation}
\usage{
lexmodelbuildingservice_get_import(importId)
}
\arguments{
\item{importId}{[required] The identifier of the import job information to return.}
}
\description{
Gets information about an import job started with the \code{StartImport}
operation.
}
\section{Request syntax}{
\preformatted{lexmodelbuildingservice$get_import(
  importId = "string"
)
}
}

\keyword{internal}
