% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rekognition_operations.R
\name{rekognition_list_stream_processors}
\alias{rekognition_list_stream_processors}
\title{Gets a list of stream processors that you have created with
CreateStreamProcessor}
\usage{
rekognition_list_stream_processors(NextToken, MaxResults)
}
\arguments{
\item{NextToken}{If the previous response was incomplete (because there are more stream
processors to retrieve), Amazon Rekognition Video returns a pagination
token in the response. You can use this pagination token to retrieve the
next set of stream processors.}

\item{MaxResults}{Maximum number of stream processors you want Amazon Rekognition Video to
return in the response. The default is 1000.}
}
\description{
Gets a list of stream processors that you have created with
CreateStreamProcessor.
}
\section{Request syntax}{
\preformatted{svc$list_stream_processors(
  NextToken = "string",
  MaxResults = 123
)
}
}

\keyword{internal}
