% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagemaker_operations.R
\name{sagemaker_describe_model}
\alias{sagemaker_describe_model}
\title{Describes a model that you created using the CreateModel API}
\usage{
sagemaker_describe_model(ModelName)
}
\arguments{
\item{ModelName}{[required] The name of the model.}
}
\description{
Describes a model that you created using the \code{CreateModel} API.
}
\section{Request syntax}{
\preformatted{svc$describe_model(
  ModelName = "string"
)
}
}

\keyword{internal}
