% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textract_operations.R
\name{textract_start_document_analysis}
\alias{textract_start_document_analysis}
\title{Starts asynchronous analysis of text for relationships in the text and
tables that are detected in a document}
\usage{
textract_start_document_analysis(DocumentLocation, FeatureTypes,
  ClientRequestToken, JobTag, NotificationChannel)
}
\arguments{
\item{DocumentLocation}{[required] The location of the document to be processed.}

\item{FeatureTypes}{[required] A list of the types of analysis to perform. Add TABLES to the list to
return information about the tables that are detected in the input
document. Add FORMS to return detected fields and the associated text.
To perform both types of analysis, add TABLES and FORMS to
\code{FeatureTypes}.}

\item{ClientRequestToken}{The idempotent token that you use to identify the start request. If you
use the same token with multiple \code{StartDocumentAnalysis} requests, the
same \code{JobId} is returned. Use \code{ClientRequestToken} to prevent the same
job from being accidentally started more than once.}

\item{JobTag}{The unique identifier you specify to identify the job in the completion
status that's published to the Amazon SNS topic.}

\item{NotificationChannel}{The Amazon SNS topic ARN that you want Amazon Textract to publish the
completion status of the operation to.}
}
\description{
Starts asynchronous analysis of text for relationships in the text and
tables that are detected in a document. Amazon Textract returns for two
types of information:
}
\details{
\itemize{
\item Words and lines that are related to nearby lines and words. The
related information is returned in two Block objects: A KEY Block
object and a VALUE Block object. For example, \emph{Name: Ana Silva
Carolina} contains a key and value. \emph{Name:} is the key. \emph{Ana Silva
Carolina} is the value.
\item Table and table cell data. A TABLE block contains information about
a detected table. A CELL block is returned for each cell in a table.
}

Amazon Textract can analyze text in document images and PDF files that
are stored in an Amazon S3 bucket. Use DocumentLocation to specify the
bucket name and file name of the document image.

\code{StartDocumentAnalysis} returns a job identifier (\code{JobId}) that you use
to get the results of the operation. When text analysis is finished,
Amazon Textract publishes a completion status to the Amazon Simple
Notification Service (Amazon SNS) topic that you specify in
\code{NotificationChannel}. To get the results of the text analysis
operation, first check that the status value published to the Amazon SNS
topic is \code{SUCCEEDED}. If so, call GetDocumentAnalysis, and pass the job
identifier (\code{JobId}) from the initial call to \code{StartDocumentAnalysis}.
}
\section{Request syntax}{
\preformatted{textract$start_document_analysis(
  DocumentLocation = list(
    S3Object = list(
      Bucket = "string",
      Name = "string",
      Version = "string"
    )
  ),
  FeatureTypes = list(
    "TABLES"|"FORMS"
  ),
  ClientRequestToken = "string",
  JobTag = "string",
  NotificationChannel = list(
    SNSTopicArn = "string",
    RoleArn = "string"
  )
)
}
}

\keyword{internal}
